moved scoring types to separate module
