+++ sciencebeam_judge/evaluation/math.py
@@ -1,0 +1,5 @@
+def mean(data):
+  return sum(data) / len(data)
+
+def safe_mean(data, default_value=0):
+  return mean(data) if data else default_value
